[FEATURE] Subscribe to jobs building cookbook dependencies
This is the first part of chained pipelines for cookbook
dependencies.

This change makes cookbook jobs to subscribe to the upstream
jobs building the `master` branch of cookbooks on which they
depend on.

Next steps should include to run `berks update` to actually
incorporate the newly uploaded cookbook versions (and commit
this back to git).