[TASK] Switch to kitchen-dokken
We (the CI) now don't put the `.kitchen.docker.yml` file into the
cookbook anymore. Instead, we check for existance and fail otherwise.

The reason for this decision is that we have to repeat the `platforms`
part of the `.kitchen.yml` again, but with docker image names given.

This is a bit sad, but it otherwise fails in `parse_image_name` split.